app/requirements.txt
Flask==2.1.0
redis==4.3.1

google-cloud-firestore==2.5.0
google-cloud-profiler==3.0.8
google-cloud-trace==1.6.1

opentelemetry-api==1.11.1
opentelemetry-exporter-gcp-trace==1.3.0
opentelemetry-sdk==1.11.1
